[Java] Modif dans le calcul

+++ java/src/util/PointXY.java
@@ -1,10 +1,10 @@
 package util;
 
 public class PointXY {
-	public float X;
+	public int X;
 	public float Y;
 	
-	public PointXY(float new_X, float new_Y) {
+	public PointXY(int new_X, float new_Y) {
 		X = new_X;
 		Y = new_Y;
 	}
